pyproject.toml
[project]
name = "contact-tracing"
description = "GraphQL API to check if a user has been exposed to COVID-19."
authors = [
    {name = "Tanmay Kacker", email = "tanmaykacker40@gmail.com"},
]
dynamic = ["version"]

dependencies = [
    "fastapi>=0.73.0",
    "strawberry-graphql[fastapi]>=0.95.1",
    "uvicorn[standard]>=0.17.1",
    "haversine>=2.5.1",
    "pytest-mock>=3.7.0",
    "types-python-dateutil>=2.8.9",
]
requires-python = ">=3.9"
license = {text = "MIT"}

[project.scripts]
dev-server = "contact_tracing.__main__:main"

[project.urls]
homepage = ""

[project.optional-dependencies]
[tool]
[tool.pdm]
version = {use_scm = true}
package-dir = "src"

[tool.pdm.dev-dependencies]
duty = ["duty>=0.7"]

format = [
    "autoflake>=1.4",
    "black>=21.10b0",
    "isort>=5.10",
]

quality = [
    "darglint>=1.8",
    "flake8-bandit>=2.1",
    "flake8-black>=0.2",
    "flake8-bugbear>=21.9",
    "flake8-builtins>=1.5",
    "flake8-comprehensions>=3.7",
    "flake8-docstrings>=1.6",
    "flake8-pytest-style>=1.5",
    "flake8-string-format>=0.3",
    "flake8-tidy-imports>=4.5",
    "flake8-variables-names>=0.0",
    "pep8-naming>=0.12",
    "wps-light>=0.15",
]
tests = [
    "pytest>=6.2",
    "pytest-cov>=3.0",
    "pytest-randomly>=3.10",
    "pytest-sugar>=0.9",
    "pytest-xdist>=2.4",
]
typing = [
    "mypy>=0.910",
    "types-markdown>=3.3",
    "types-toml>=0.10",
]

[tool.black]
line-length = 120
exclude = "tests/fixtures"

[tool.isort]
line_length = 120
not_skip = "__init__.py"
multi_line_output = 3
force_single_line = false
balanced_wrapping = true
default_section = "THIRDPARTY"
known_first_party = "contact-tracing"
include_trailing_comma = true


[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"
